+++ packages/components/src/index.ts
@@ -7,3 +7,4 @@ export { MinisButton } from './components/button/button.js';
 export { MinisPillCounter } from './components/pill-counter/pill-counter.js';
 export { MinisAlert } from './components/alert/alert.js';
 export { MinisMessage } from './components/message/message.js';
+export { MinisTag } from './components/tag/tag.js';